caches: don't try to delete keys if there aren't any to delete.
- This is problematic for REDIS backend.

+++ rhodecode/lib/rc_cache/utils.py
@@ -187,8 +187,10 @@ def get_or_create_region(region_name, region_namespace=None):
 def clear_cache_namespace(cache_region, cache_namespace_uid):
     region = get_or_create_region(cache_region, cache_namespace_uid)
     cache_keys = region.backend.list_keys(prefix=cache_namespace_uid)
-    region.delete_multi(cache_keys)
-    return len(cache_keys)
+    num_delete_keys = len(cache_keys)
+    if num_delete_keys:
+        region.delete_multi(cache_keys)
+    return num_delete_keys
 
 
 class ActiveRegionCache(object):
